Refactor component imports to new base structure
- Updated import paths for various components to align with the new directory structure under `ui/components/base/`.
- Refactored components including AnalyticsDashboard, AnimeCard, BatchDownloadProgress, and others to utilize the new base imports.
- Removed obsolete UI component files that have been migrated to the new structure, enhancing maintainability and organization.
