+++ src/components/ErrorBoundary.tsx
@@ -1,6 +1,6 @@
 import { Component, type ErrorInfo, type ReactNode } from "react";
 import { AlertTriangle } from "lucide-react";
-import { Button } from "./ui/button";
+import { Button } from "../ui/components/base/button";
 import posthog from "posthog-js";
 
 interface Props {
