+++ src/components/LibraryView.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
-import { Input } from "./ui/input";
-import { Button } from "./ui/button";
-import { Card, CardContent } from "./ui/card";
+import { Input } from "../ui/components/base/input";
+import { Button } from "../ui/components/base/button";
+import { Card, CardContent } from "../ui/components/base/card";
 import { Search, Download, Upload, HardDrive, Film } from "lucide-react";
 import { AnimeStats, LibraryStats } from "../core/types";
 import {
